+++ swov1/RELEASES.md
@@ -45,3 +45,13 @@ Based on:
 - [typescript v0.5.0] swov1
 ### Releases
 - [NPM v0.5.0] https://www.npmjs.com/package/@solarwinds/swo-sdk-typescript/v/0.5.0 - swov1
+
+## 2025-11-14 00:03:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.656.2 (2.753.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0] swov1
+### Releases
+- [NPM v0.6.0] https://www.npmjs.com/package/@solarwinds/swo-sdk-typescript/v/0.6.0 - swov1